refactor(*): tidying up
