+++ server/src/components.js
@@ -1,9 +1,9 @@
-const connectToMongoDB = require("./core/connectToMongoDB");
-const createMailer = require("./core/mailer");
-const creatHttpClient = require("./core/httpClient");
+const connectToMongoDB = require("./modules/connectToMongoDB");
+const createMailer = require("./modules/mailer");
+const creatHttpClient = require("./modules/httpClient");
 const campagnesDAO = require("./dao/campagnes.dao");
 const temoignagesDAO = require("./dao/temoignages.dao.js");
-const createLogger = require("./core/logger");
+const createLogger = require("./modules/logger");
 const usersDAO = require("./dao/users.dao");
 const defaults = require("./config");
 const Log = require("./models/log");
