+++ server/src/utils/authenticate.utils.js
@@ -1,5 +1,5 @@
 const jwt = require("jsonwebtoken");
-const config = require("../../config");
+const config = require("../config");
 
 const isDev = config.env === "dev";
